Seperate UI into own class
